refactor: :truck: utils.jsをshred/utilsディレクトリに移動させてリネーム
