+++ packages/app/src/components/buttons/GameResetButton.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/components/ui/Tooltip";
 import { useResetGame } from "@/game/hooks/useResetGame";
 import { useScene } from "@/game/hooks/useScene";
-import { preloadImages } from "@/lib/utils";
+import { preloadImages } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { HelpOverlay } from "../ui/HelpOverlay";
