+++ packages/app/src/components/buttons/HelpButton.tsx
@@ -6,7 +6,7 @@ import {
   TooltipTrigger,
 } from "@/components/ui/Tooltip";
 import { useHelp } from "@/game/hooks/useHelp";
-import { preloadImages } from "@/lib/utils";
+import { preloadImages } from "@/shared/utils/helpers";
 import { Tooltip } from "@radix-ui/react-tooltip";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
