+++ packages/app/src/components/scenes/ModeSelectScene.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { MultiPlayButton } from "../buttons/MultiPlayButton";
