+++ packages/app/src/components/texts/GameTurnText.tsx
@@ -1,6 +1,6 @@
 import { NeonText } from "@/components/ui/NeonText";
 import { useGame } from "@/game/hooks/useGame";
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
