+++ packages/app/src/components/ui/AppBackground.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 
 export interface AppBackgroundProps {
