+++ packages/app/src/components/ui/ImageButton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 import { forwardRef } from "react";
 import { Button, type ButtonProps } from "./Button";
