+++ packages/app/src/components/ui/Tooltip.tsx
@@ -1,7 +1,7 @@
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 
 const TooltipProvider = TooltipPrimitive.Provider;
